For Compilations with Visual Studio for Windows.

+++ source/search.cpp
@@ -26,7 +26,7 @@
 #include <cassert>
 #include <cmath>
 #include <cstring>   // For std::memset
-#include <unistd.h> //for sleep
+//#include <unistd.h> //for sleep
 #include <iostream>
 #include <sstream>
 #include <random>
@@ -364,7 +364,7 @@ void Thread::search() {
 
   size_t multiPV = Options["MultiPV"];
   Skill skill(Options["Skill Level"]);
-  if (tactical) multiPV = pow(2, tactical);
+  if (tactical) multiPV = size_t(pow(2, tactical));
   
   // When playing with strength handicap enable MultiPV search that we will
   // use behind the scenes to retrieve a set of possible moves.
